# GitMiner/Documentation/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /gitminer/issues/{id}:
    get:
      tags:
      - Issue
      - issue
      - get
      summary: Retrieve an issue
      description: Get an issue
      operationId: findOne_2
      parameters:
      - name: id
        in: path
        description: Id of the issue
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        "404":
          description: Not Found
    put:
      tags:
      - Issue
      - issue
      - update
      summary: Update an issue
      description: Update an existing issue with data passed on the body of the request
      operationId: update_2
      parameters:
      - name: id
        in: path
        description: Id of the issue to update
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Issue'
        required: true
      responses:
        "400":
          description: Bad Request
        "204":
          description: No Content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        "404":
          description: Not Found
    delete:
      tags:
      - Issue
      - issue
      - delete
      summary: Delete an issue
      description: Delete an existing issue by the id specified
      operationId: delete_2
      parameters:
      - name: id
        in: path
        description: Issue id to be deleted
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
        "204":
          description: No Content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        "404":
          description: Not Found
...
  /gitminer/issues:
    get:
      tags:
      - Issue
      - get
      - issues
      summary: Retrieve a list of issues
      description: Get a list of issues
      operationId: findIssues
      parameters:
      - name: authorId
        in: query
        description: User name of the author
        required: false
        schema:
          type: string
      - name: state
        in: query
        description: State of the issue
        required: false
        schema:
          type: string
      - name: keyword
        in: query
        description: Keyword to filter
        required: false
        schema:
          type: string
      - name: page
        in: query
        description: Page retrieved
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: order
        in: query
        required: false
        schema:
          type: string
      - name: size
        in: query
        description: Number of elements retrieved
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        "404":
          description: Not Found
    post:
      tags:
      - Issue
      - issue
      - create
      summary: Create an issue
      description: Create an issue whose data is passed in the body of the request
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Issue'
        required: true
      responses:
        "400":
          description: Bad Request
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
...
  /gitminer/issues/{id}/comments:
    get:
      tags:
      - comments
      - Issue
      - issue
      - get
      summary: Retrieve comments of an issue
      description: Get comments of an issue
      operationId: findCommentsOfOne
      parameters:
      - name: id
        in: path
        description: Id of the issue
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        "404":
          description: Not Found
components:
  schemas:
    User:
      required:
      - username
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        avatar_url:
          type: string
        web_url:
          type: string
    Comment:
      required:
      - body
      - created_at
      type: object
      properties:
        id:
          type: string
        body:
          type: string
        author:
          $ref: '#/components/schemas/User'
        created_at:
          type: string
        updated_at:
          type: string
...
    Issue:
      type: object
      properties:
        id:
          type: string
        ref_id:
          type: string
        title:
          type: string
        description:
          type: string
        state:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
        closed_at:
          type: string
        labels:
          type: array
          items:
            type: string
        author:
          $ref: '#/components/schemas/User'
        assignee:
          $ref: '#/components/schemas/User'
        upvotes:
          type: integer
          format: int32
        downvotes:
          type: integer
          format: int32
        web_url:
          type: string
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
...